+++ addons/discord_gd/classes/message_button.gd
@@ -1,6 +1,9 @@
 class_name MessageButton
+"""
+Represents a Discord message button.
+"""
 
-var _STYLES = {0: "DEFAULT", 1: "PRIMARY", 2: "SECONDARY", 3: "SUCCESS", 4: "DANGER", 5: "LINK"}
+var _STYLES = {0: 'DEFAULT', 1: 'PRIMARY', 2: 'SECONDARY', 3: 'SUCCESS', 4: 'DANGER', 5: 'LINK'}
 
 enum STYLES { DEFAULT, PRIMARY, SECONDARY, SUCCESS, DANGER, LINK }
 
@@ -88,10 +91,10 @@ func _to_dict() -> Dictionary:
 
 	if _style == STYLES.LINK:
 		# Must have a url
-		assert(Helpers.is_valid_str(url), "A LINK MessageButton must have a url.")
+		assert(Helpers.is_valid_str(url), 'A LINK MessageButton must have a url.')
 		return {'type': type, 'style': _style, 'label': label, 'url': url, 'disabled': disabled}
 	else:
-		assert(Helpers.is_valid_str(custom_id), "A button must have a custom_id.")
+		assert(Helpers.is_valid_str(custom_id), 'A button must have a custom_id.')
 		return {
 			'type': type,
 			'style': _style,