+++ addons/discord_gd/classes/select_menu.gd
@@ -1,4 +1,7 @@
 class_name SelectMenu
+"""
+Represents a Discord select menu.
+"""
 
 var custom_id: String setget set_custom_id, get_custom_id
 var placeholder: String setget set_placeholder, get_placeholder
@@ -13,7 +16,7 @@ var type: int = 3
 
 
 func set_custom_id(new_custom_id):
-	Helpers.assert_length(new_custom_id, 100, "custom_id of SelectMenu cannot be more than 100 characters.")
+	Helpers.assert_length(new_custom_id, 100, 'custom_id of SelectMenu cannot be more than 100 characters.')
 	custom_id = new_custom_id
 	return self
 
@@ -22,30 +25,30 @@ func get_custom_id() -> String:
 
 
 func add_option(value: String, label: String, data: Dictionary = {}):
-	assert(options.size() <= 25, "options ofSelectMenu cannot have more than 25 options")
-	assert(Helpers.is_valid_str(value), "value of SelectMenu option  must be a valid String")
-	Helpers.assert_length(value, 100, "value of SelectMenu option cannot be more than 100 characters")
-	assert(Helpers.is_valid_str(label), "SelectMenu option must have a label")
-	Helpers.assert_length(label, 100, "label of SelectMenu option cannot be more than 100 characters")
+	assert(options.size() <= 25, 'options of SelectMenu cannot have more than 25 options')
+	assert(Helpers.is_valid_str(value), 'value of SelectMenu option  must be a valid String')
+	Helpers.assert_length(value, 100, 'value of SelectMenu option cannot be more than 100 characters')
+	assert(Helpers.is_valid_str(label), 'SelectMenu option must have a label')
+	Helpers.assert_length(label, 100, 'label of SelectMenu option cannot be more than 100 characters')
 
 	# Parse data
 	#{description: "", emoji: {}, default = false}
 	var _data = {
-		"value": value,
-		"label": label
+		'value': value,
+		'label': label
 	}
-	if data.has("description"):
-		assert(typeof(data.description) == TYPE_STRING, "description of SelectMenu option must be a String")
-		Helpers.assert_length(data.description, 100, "description of SelectMenu cannot be more than 100 characters")
-		_data["description"] = data.description
+	if data.has('description'):
+		assert(typeof(data.description) == TYPE_STRING, 'description of SelectMenu option must be a String')
+		Helpers.assert_length(data.description, 100, 'description of SelectMenu cannot be more than 100 characters')
+		_data['description'] = data.description
 
-	if data.has("emoji"):
-		assert(typeof(data.emoji) == TYPE_DICTIONARY, "emoji of SelecMenu option must be a Dictionary")
-		_data["emoji"] = data.emoji
+	if data.has('emoji'):
+		assert(typeof(data.emoji) == TYPE_DICTIONARY, 'emoji of SelecMenu option must be a Dictionary')
+		_data['emoji'] = data.emoji
 
-	if data.has("default"):
-		assert(typeof(data.default) == TYPE_BOOL, "default of SelectMenu option must be a bool")
-		_data["default"] = data.default
+	if data.has('default'):
+		assert(typeof(data.default) == TYPE_BOOL, 'default of SelectMenu option must be a bool')
+		_data['default'] = data.default
 
 	options.append(_data)
 
@@ -60,7 +63,7 @@ func get_options() -> Array:
 
 
 func set_placeholder(new_placeholder: String):
-	Helpers.assert_length(new_placeholder, 100, "placeholder of SelectMenu cannot be more than 100 characters.")
+	Helpers.assert_length(new_placeholder, 100, 'placeholder of SelectMenu cannot be more than 100 characters.')
 	placeholder = new_placeholder
 	return self
 
@@ -69,7 +72,7 @@ func get_placeholder() -> String:
 
 
 func set_min_values(new_min_values: int):
-	assert(new_min_values <= 25, "min_values of SelectMenu cannot be more than 25")
+	assert(new_min_values <= 25, 'min_values of SelectMenu cannot be more than 25')
 	min_values = new_min_values
 	return self
 
@@ -78,7 +81,7 @@ func get_min_values() -> int:
 
 
 func set_max_values(new_max_values: int):
-	assert(new_max_values <= 25, "max_values of SelectMenu cannot be more than 25")
+	assert(new_max_values <= 25, 'max_values of SelectMenu cannot be more than 25')
 	max_values = new_max_values
 	return self
 
@@ -99,20 +102,20 @@ func _init():
 	return self
 
 func _to_string(pretty: bool = false) -> String:
-	return JSON.print(_to_dict(), "\t") if pretty else JSON.print(_to_dict())
+	return JSON.print(_to_dict(), '\t') if pretty else JSON.print(_to_dict())
 
 func print():
 	print(_to_string(true))
 
 func _to_dict() -> Dictionary:
 	# Default style is primary
-	assert(Helpers.is_valid_str(custom_id), "A button must have a custom_id.")
+	assert(Helpers.is_valid_str(custom_id), 'A button must have a custom_id.')
 	return {
-		"type": type,
-		"custom_id": custom_id,
-		"options": options,
-		"placeholder": placeholder,
-		"min_values": min_values,
-		"max_values": max_values,
-		"disabled": disabled,
+		'type': type,
+		'custom_id': custom_id,
+		'options': options,
+		'placeholder': placeholder,
+		'min_values': min_values,
+		'max_values': max_values,
+		'disabled': disabled,
 	}